Add files via upload
The dataset (after some pre-processing) had 905 users with 207699 books. Applied feedback alignment on a book recommendation problem. The network consisted of 3 layers with 200 hidden units, batch size of 50 and iterations=1000.  To increase speed the dimensions of the ratings matrix was reduced to 905 X 50. We fed the ratings matrix(905 X 50) as the input to the matrix and tried to overfit the output to this matrix. 
The results were unstable for feedback alignment.